updates types declaration and build
